requirements.txt
streamlit
groq
PyPDF2
python-dotenv
matplotlib
reportlab